+++ fitgrid/__init__.py
@@ -1,5 +1,6 @@
 from .fake_data import generate
 from .io import epochs_from_hdf, epochs_from_dataframe, load_grid
+from .models import lm, lmer
 
 #: default epoch identifier
 EPOCH_ID = 'Epoch_idx'
